LOG-3832: fix tests to support different base collector images
