+++ test/e2e/collection/suite_test.go
@@ -1,0 +1,13 @@
+package collection
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestSuite(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "ClusterLogging E2E Suite - Collection")
+}
